Split reconfig integration tests
