+++ test/reconfig/configtx/config_disseminate_test.go
@@ -4,7 +4,7 @@ Copyright IBM Corp. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package reconfig
+package configtx
 
 import (
 	"context"
