+++ test/reconfig/endpoints/config_endpoint_update_test.go
@@ -1,0 +1,279 @@
+/*
+Copyright IBM Corp. All Rights Reserved.
+
+SPDX-License-Identifier: Apache-2.0
+*/
+
+package endpoints
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+	"strconv"
+	"strings"
+	"sync/atomic"
+	"testing"
+	"time"
+
+	"github.com/hyperledger/fabric-protos-go-apiv2/common"
+	"github.com/hyperledger/fabric-x-common/protoutil"
+	"github.com/hyperledger/fabric-x-common/tools/configtxgen"
+	"github.com/hyperledger/fabric-x-orderer/common/tools/armageddon"
+	"github.com/hyperledger/fabric-x-orderer/common/types"
+	"github.com/hyperledger/fabric-x-orderer/common/utils"
+	"github.com/hyperledger/fabric-x-orderer/config"
+	test_utils "github.com/hyperledger/fabric-x-orderer/test/utils"
+	"github.com/hyperledger/fabric-x-orderer/testutil"
+	"github.com/hyperledger/fabric-x-orderer/testutil/client"
+	"github.com/hyperledger/fabric-x-orderer/testutil/configutil"
+	"github.com/hyperledger/fabric-x-orderer/testutil/signutil"
+	"github.com/hyperledger/fabric-x-orderer/testutil/tx"
+	"github.com/onsi/gomega/gexec"
+	"github.com/stretchr/testify/require"
+	"go.uber.org/zap"
+)
+
+var caFolders = map[string]struct{}{
+	"ca":         {},
+	"tlsca":      {},
+	"cacerts":    {},
+	"tlscacerts": {},
+}
+
+func TestUpdatePartyRouterEndpoint(t *testing.T) {
+	// Prepare Arma config and crypto and get the genesis block
+	dir, err := os.MkdirTemp("", t.Name())
+	require.NoError(t, err)
+	defer os.RemoveAll(dir)
+
+	configPath := filepath.Join(dir, "config.yaml")
+	numOfParties := 4
+	submittingParty := types.PartyID(1)
+
+	netInfo := testutil.CreateNetwork(t, configPath, numOfParties, 2, "none", "none")
+	defer netInfo.CleanUp()
+	require.NotNil(t, netInfo)
+
+	armageddon.NewCLI().Run([]string{"generate", "--config", configPath, "--output", dir})
+
+	armaBinaryPath, err := gexec.BuildWithEnvironment("github.com/hyperledger/fabric-x-orderer/cmd/arma", []string{"GOPRIVATE=" + os.Getenv("GOPRIVATE")})
+	defer gexec.CleanupBuildArtifacts()
+	require.NoError(t, err)
+	require.NotNil(t, armaBinaryPath)
+
+	// Start Arma nodes
+	numOfArmaNodes := len(netInfo)
+	readyChan := make(chan string, numOfArmaNodes)
+	armaNetwork := testutil.RunArmaNodes(t, dir, armaBinaryPath, readyChan, netInfo)
+	defer armaNetwork.Stop()
+
+	testutil.WaitReady(t, readyChan, numOfArmaNodes, 10)
+
+	userConfig, err := testutil.GetUserConfig(dir, submittingParty)
+	require.NoError(t, err)
+	require.NotNil(t, userConfig)
+
+	totalTxNumber := 100
+
+	// rate limiter parameters
+	fillInterval := 10 * time.Millisecond
+	fillFrequency := 1000 / int(fillInterval.Milliseconds())
+	rate := 500
+	capacity := rate / fillFrequency
+	rl, err := armageddon.NewRateLimiter(rate, fillInterval, capacity)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "failed to start a rate limiter")
+		os.Exit(3)
+	}
+
+	broadcastClient := client.NewBroadcastTxClient(userConfig, 10*time.Second)
+	signer, certBytes, err := testutil.LoadCryptoMaterialsFromDir(t, userConfig.MSPDir)
+	require.NoError(t, err)
+
+	org := fmt.Sprintf("org%d", submittingParty)
+
+	for i := range totalTxNumber {
+		status := rl.GetToken()
+		if !status {
+			fmt.Fprintf(os.Stderr, "failed to send tx %d", i+1)
+			os.Exit(3)
+		}
+		txContent := tx.PrepareTxWithTimestamp(i, 64, []byte("sessionNumber"))
+		env := tx.CreateSignedStructuredEnvelope(txContent, signer, certBytes, org)
+		err = broadcastClient.SendTx(env)
+		require.NoError(t, err)
+	}
+
+	var parties []types.PartyID
+	for i := 1; i <= numOfParties; i++ {
+		parties = append(parties, types.PartyID(i))
+	}
+
+	pullRequestSigner := signutil.CreateTestSigner(t, "org1", dir)
+
+	statusUknown := common.Status_UNKNOWN
+	test_utils.PullFromAssemblers(t, &test_utils.BlockPullerOptions{
+		UserConfig:   userConfig,
+		Parties:      parties,
+		Transactions: totalTxNumber,
+		Timeout:      60,
+		ErrString:    "cancelled pull from assembler: %d; pull ended: failed to receive a deliver response: rpc error: code = Canceled desc = grpc: the client connection is closing",
+		Status:       &statusUknown,
+		Signer:       pullRequestSigner,
+	})
+
+	// Create config update
+	configUpdateBuilder, cleanUp := configutil.NewConfigUpdateBuilder(t, dir, filepath.Join(dir, "bootstrap", "bootstrap.block"))
+	defer cleanUp()
+
+	partyToUpdate := submittingParty
+	nonUpdatedRouterParties := []types.PartyID{2, 3, 4}
+	routerIP := strings.Split(userConfig.RouterEndpoints[partyToUpdate-1], ":")[0] // extract IP from the user config router endpoint
+	availablePort, newListener := testutil.SharedTestPortAllocator().Allocate(t)
+	newPort, err := strconv.Atoi(availablePort)
+	require.NoError(t, err)
+	routerToUpdate := armaNetwork.GetRouter(t, submittingParty)
+	// Close the previous listener before replacing it
+	if routerToUpdate.Listener != nil {
+		routerToUpdate.Listener.Close()
+	}
+	routerToUpdate.Listener = newListener
+
+	configUpdatePbData := configUpdateBuilder.UpdateRouterEndpoint(t, partyToUpdate, routerIP, newPort)
+
+	// Create config tx
+	env := configutil.CreateConfigTX(t, dir, parties, int(submittingParty), configUpdatePbData)
+	require.NotNil(t, env)
+
+	// Send the config tx
+	err = broadcastClient.SendTxTo(env, submittingParty)
+	require.NoError(t, err)
+
+	broadcastClient.Stop()
+
+	t.Log("Wait for the router to enter pending admin state and then stop it")
+	testutil.WaitForPendingAdminByTypeAndParty(t, netInfo, []testutil.NodeType{testutil.Router}, []types.PartyID{partyToUpdate})
+	armaNetwork.GetRouter(t, partyToUpdate).StopArmaNode()
+
+	t.Log("Wait for arma nodes to restart dynamically")
+	testutil.WaitForRelaunchByType(t, netInfo, []testutil.NodeType{testutil.Consensus, testutil.Assembler, testutil.Batcher}, 1)
+	testutil.WaitForRelaunchByTypeAndParty(t, netInfo, []testutil.NodeType{testutil.Router}, nonUpdatedRouterParties, 1)
+
+	// Pull blocks to verify all transactions are included
+	userBlockHandler := &verifyRouterEndpointUpdate{updatedParty: partyToUpdate, routerIP: routerIP, newPort: newPort}
+	test_utils.PullFromAssemblers(t, &test_utils.BlockPullerOptions{
+		UserConfig:   userConfig,
+		Parties:      parties,
+		Transactions: totalTxNumber + 1, // including config update tx
+		Timeout:      60,
+		BlockHandler: userBlockHandler,
+		ErrString:    "cancelled pull from assembler: %d; pull ended: failed to receive a deliver response: rpc error: code = Canceled desc = grpc: the client connection is closing",
+		Status:       &statusUknown,
+		Signer:       pullRequestSigner,
+	})
+
+	require.True(t, userBlockHandler.RouterEndpointUpdated.Load(), "Router endpoint was not updated in the config update")
+
+	// Verify the config stored in the router's config store is updated
+	routerNodeConfigPath := filepath.Join(dir, "config", fmt.Sprintf("party%d", partyToUpdate), "local_config_router.yaml")
+	cfg, _, err := config.ReadConfig(routerNodeConfigPath, testutil.CreateLoggerForModule(t, "ReadConfigRouter", zap.DebugLevel))
+	require.NoError(t, err)
+	require.True(t, cfg.SharedConfig.GetPartiesConfig()[partyToUpdate-1].RouterConfig.Host == routerIP &&
+		cfg.SharedConfig.GetPartiesConfig()[partyToUpdate-1].RouterConfig.Port == uint32(newPort), "Shared config was not updated with the new router endpoint")
+
+	// Update the router node local config with the new endpoint to allow it to start
+	localConfig, _, err := config.LoadLocalConfig(routerNodeConfigPath)
+	require.NoError(t, err)
+	localConfig.NodeLocalConfig.GeneralConfig.ListenAddress = routerIP
+	localConfig.NodeLocalConfig.GeneralConfig.ListenPort = uint32(newPort)
+	utils.WriteToYAML(localConfig.NodeLocalConfig, routerNodeConfigPath)
+
+	t.Log("Restart Router")
+	armaNetwork.GetRouter(t, partyToUpdate).RestartArmaNode(t, readyChan)
+
+	testutil.WaitReady(t, readyChan, 1, 10)
+
+	// Update the user config with the new router endpoint
+	userConfig.RouterEndpoints[partyToUpdate-1] = fmt.Sprintf("%s:%d", routerIP, newPort)
+	broadcastClient = client.NewBroadcastTxClient(userConfig, 10*time.Second)
+
+	t.Log("Send transactions again and verify they are processed")
+	for i := range totalTxNumber {
+		status := rl.GetToken()
+		if !status {
+			fmt.Fprintf(os.Stderr, "failed to send tx %d", i+1)
+			os.Exit(3)
+		}
+		txContent := tx.PrepareTxWithTimestamp(i+totalTxNumber, 64, []byte("sessionNumber"))
+		env := tx.CreateSignedStructuredEnvelope(txContent, signer, certBytes, org)
+		err = broadcastClient.SendTx(env)
+		require.NoError(t, err)
+	}
+
+	broadcastClient.Stop()
+
+	// Pull blocks to verify all transactions are included
+	userBlockHandler = &verifyRouterEndpointUpdate{updatedParty: partyToUpdate, routerIP: routerIP, newPort: newPort}
+	test_utils.PullFromAssemblers(t, &test_utils.BlockPullerOptions{
+		UserConfig:   userConfig,
+		Parties:      parties,
+		Transactions: totalTxNumber*2 + 1, // including config update tx
+		Timeout:      60,
+		BlockHandler: userBlockHandler,
+		ErrString:    "cancelled pull from assembler: %d; pull ended: failed to receive a deliver response: rpc error: code = Canceled desc = grpc: the client connection is closing",
+		Status:       &statusUknown,
+		Signer:       pullRequestSigner,
+	})
+
+	require.True(t, userBlockHandler.RouterEndpointUpdated.Load(), "Router endpoint was not updated in the config update")
+	armaNetwork.Stop()
+}
+
+// Verify that the config update is applied by checking the router endpoint in the config update block
+type verifyRouterEndpointUpdate struct {
+	updatedParty          types.PartyID
+	RouterEndpointUpdated atomic.Bool
+	routerIP              string
+	newPort               int
+}
+
+func (v *verifyRouterEndpointUpdate) HandleBlock(t *testing.T, block *common.Block) error {
+	if protoutil.IsConfigBlock(block) {
+		envelope, err := configutil.ReadConfigEnvelopeFromConfigBlock(block)
+		if err != nil || envelope == nil {
+			return fmt.Errorf("failed to read config envelope from config block: %w", err)
+		}
+
+		partyConfig := configutil.GetPartyConfig(t, envelope, v.updatedParty)
+		if partyConfig == nil {
+			return fmt.Errorf("party config for party %d not found in the config block", v.updatedParty)
+		}
+
+		v.RouterEndpointUpdated.Store(partyConfig.RouterConfig.Host == v.routerIP && partyConfig.RouterConfig.Port == uint32(v.newPort))
+	}
+
+	return nil
+}
+
+type exportConfigBlockToFile struct {
+	configSeq uint64
+	path      string
+}
+
+func (ec *exportConfigBlockToFile) HandleBlock(t *testing.T, block *common.Block) error {
+	if protoutil.IsConfigBlock(block) {
+		env, err := protoutil.ExtractEnvelope(block, 0)
+		require.NoError(t, err)
+		payload, err := protoutil.UnmarshalPayload(env.Payload)
+		require.NoError(t, err)
+		configEnv, err := protoutil.UnmarshalConfigEnvelope(payload.Data)
+		require.NoError(t, err)
+		if configEnv.GetConfig().GetSequence() == ec.configSeq {
+			configBlock := &common.Block{Header: block.GetHeader(), Data: block.GetData(), Metadata: block.GetMetadata()}
+			err := configtxgen.WriteOutputBlock(configBlock, ec.path)
+			require.NoError(t, err)
+		}
+	}
+	return nil
+}
